cmd,pkg,test: Move the deploy package down a directory level.
