+++ pkg/deploy/uninstall.go
