+++ test/deployframework/framework_test.go
@@ -8,7 +8,7 @@ import (
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	metering "github.com/operator-framework/operator-metering/pkg/apis/metering/v1"
-	"github.com/operator-framework/operator-metering/pkg/operator/deploy"
+	"github.com/operator-framework/operator-metering/pkg/deploy"
 )
 
 func TestNewDeployerConfig(t *testing.T) {
